Update draw.js
Second version from CodePen, slightly refactored. I simplified some of the argument options and got rid of the 'x' and 'y' types, in favor of parameter arrays. The order of the arguments matters in the transform() method, so the first argument of transform will be the varying variable (so if you wanted an x-curve, you'd pass in the x parameter first, which is the 0th item of the parameter array). The method accepts any number of arguments afterwards, and they get added to a constantParameters array. So for the structure of the for loops, as long as the varyingParameter is on the very inside, it doesn't matter what order the constant parameters are figured out in the outer for loops. I would like to generalize that section of the code for any number of parameters, which I've done for everything else, but right now it's still just x and y.